+++ src/app/app-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './home/home.component';
-import { UserHomeComponent } from './user-home/user-home.component';
-import { LoginComponent } from './login/login.component';
-import { UserListComponent } from './user-list/user-list.component';
-import { UserSignUpComponent } from './user-sign-up/user-sign-up.component';
+import { HomeComponent } from './components/home/home.component';
+import { UserHomeComponent } from './components/user-home/user-home.component';
+import { LoginComponent } from './components/login/login.component';
+import { UserListComponent } from './components/user-list/user-list.component';
+import { UserSignUpComponent } from './components/user-sign-up/user-sign-up.component';
 
 const routes: Routes = [
   { path: '', component: HomeComponent },
