+++ src/app/app.component.html
@@ -2,12 +2,7 @@
 <div class="container">
   <div class="row">
     <div class="col-md-12">
-
-      <!-- <router-outlet (activate)="getPrincipal()"></router-outlet> -->
-
-      <app-blogsview></app-blogsview>
-
-     
+      <router-outlet (activate)="getPrincipal()"></router-outlet>
     </div>
   </div>
 </div>
