+++ src/app/app.component.ts
@@ -1,19 +1,12 @@
 import { Component } from '@angular/core';
-import { User } from './user';
-import { AuthService } from './auth.service';
+import { User } from './models/user';
+import { AuthService } from './services/auth.service';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css']
 })
-// export class AppComponent {
-
-//   title: string;
-
-//   constructor() {
-//     this.title = 'Users';
-//   }
 
 export class AppComponent {
   title = 'Talks Rain or Shine!';
