--- src/app/app/app.component.html
@@ -1,8 +1,0 @@
-<app-navbar [loggedInUser]="principal"></app-navbar>
-<div class="container">
-  <div class="row">
-    <div class="col-md-12">
-      <router-outlet (activate)="getPrincipal()"></router-outlet>
-    </div>
-  </div>
-</div>
