--- src/app/app/app.module.ts
@@ -1,34 +1,0 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { AppRoutingModule } from './app-routing.module';
-import { FormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
-import { AppComponent } from './app.component';
-import { UserListComponent } from '../user-list/user-list.component';
-import { UserService } from '../user-service/user.service';
-import { UserSignUpComponent } from '../user-sign-up/user-sign-up.component';
-import { LoginComponent } from '../login/login.component';
-import { NavbarComponent } from '../navbar/navbar.component';
-import { UserHomeComponent } from '../user-home/user-home.component';
-import { HomeComponent } from '../home/home.component';
-
-@NgModule({
-  declarations: [
-    AppComponent,
-    UserListComponent,
-    UserSignUpComponent,
-    LoginComponent,
-    NavbarComponent,
-    UserHomeComponent,
-    HomeComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule,
-    FormsModule
-  ],
-  providers: [UserService],
-  bootstrap: [AppComponent]
-})
-export class AppModule { }
