+++ src/app/components/blogsview/blogsview.component.html
@@ -8,6 +8,7 @@
             <th scope="col">Body</th>
             <th scope="col">Username</th>
             <th scope="col">PostedDate</th>
+            <th *ngIf="loggedInUser.role === 1" scope="col">Delete</th>
           </tr>
         </thead>
         <tbody>
@@ -15,8 +16,8 @@
             <td>{{ blog.id }}</td>
             <td>{{ blog.body }}</td>
             <td>{{ blog.userId }}</td>
-            <td>{{ blog.timeCreated}}</td>
-
+            <td>{{ blog.timeCreated }}</td>
+            <td *ngIf="loggedInUser.role === 1" (click)="deleteBlog(blog.id)"><i class="fas fa-trash-alt"></i></td>
           </tr>
         </tbody>
       </table>
