+++ src/app/components/blogsview/blogsview.component.ts
@@ -1,8 +1,7 @@
-import { Component, OnInit } from '@angular/core';
-import { Blog } from '../blog';
-import { BlogService } from '../blog.service';
-
-
+import { Component, Input, OnInit } from '@angular/core';
+import { Blog } from 'src/app/models/blog';
+import { BlogService } from 'src/app/services/blog.service';
+import { User } from 'src/app/models/user';
 
 @Component({
   selector: 'app-blogsview',
@@ -10,7 +9,10 @@ import { BlogService } from '../blog.service';
   styleUrls: ['./blogsview.component.css']
 })
 export class BlogsviewComponent implements OnInit {
+  
   blogs:Blog[];
+
+  @Input() loggedInUser: User;
 
   constructor(private blogService:BlogService) { }
 
@@ -27,6 +29,15 @@ export class BlogsviewComponent implements OnInit {
       console.log(err); }
     );
     
+  }
+
+  deleteBlog(id: number) {
+    this.blogService.delete(id).subscribe(
+      data => {
+        console.log('deleted response', data);
+        this.getBlogs();
+      }
+    )
   }
 
 }
