+++ src/app/components/user-home/user-home.component.html
@@ -1,3 +1,4 @@
 <main class="container-fluid mx-auto " style="padding: 0 7.5em;">
     <h1 class="display-1 text-center" id="welcome">Welcome!</h1>
+    <app-blogsview [loggedInUser]="principal"></app-blogsview>
 </main>
