+++ src/app/components/user-sign-up/user-sign-up.component.css
