+++ src/app/services/auth.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { environment } from 'src/environments/environment';
-import { User } from '../user/user';
+import { User } from '../models/user';
 import { map } from 'rxjs/operators'
 
 @Injectable({
