+++ src/app/services/user.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { User } from '../user/user';
+import { User } from '../models/user';
 import { Observable} from 'rxjs';
 
 @Injectable()
