--- src/app/user.service.ts
@@ -1,22 +1,0 @@
-import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { User } from './user';
-import { Observable} from 'rxjs';
-
-@Injectable()
-export class UserService {
-
-  private usersUrl: string;
-
-  constructor(private http: HttpClient) {
-    this.usersUrl = 'http://localhost:8080/users';
-  }
-
-  public findAll(): Observable<User[]> {
-    return this.http.get<User[]>(this.usersUrl);
-  }
-
-  public save(user: User) {
-    return this.http.post<User>(this.usersUrl, user);
-  }
-}
