--- src/app/user.ts
@@ -1,10 +1,0 @@
-export class User {
-    id: string;
-    firstName: string;
-    lastName: string;
-    email: string;
-    username: string;
-    password: string;
-    location: number;
-    role: 1;
-}
