+++ src/environments/environment.ts
@@ -4,7 +4,7 @@
 
 export const environment = {
   production: true,
-  apiUrl: 'http://localhost:8081'
+  apiUrl: 'http://localhost:8080'
 };
 
 /*
